refactor: move dashboard feature to features/dashboard/
